refactor(engine): fold eq-operand rewrite into LowerStatement visitor
  Replace the standalone `RewriteEqOperand` with
  `LowerStatement::rewrite_eq_operand`
  invoked from an overridden `visit_expr_binary_op_mut`.  The lowering walk's
  existing scope management is reused, eliminating the four duplicated visit
  overrides on the standalone visitor.
